ai/requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0

